test: Move product search caching test and fix cache ref
