+++ controllers/__tests__/bookings.js
@@ -2,11 +2,9 @@
 
 const chance = require('chance').Chance();
 const hash = require('object-hash');
-// cache is now mocked at the top level
-// const cache = require('../../cache');
+const cache = require('../../cache');
 
 const { env: { adminKey } } = process;
-jest.mock('../../cache');
 
 describe('user: bookings controller', () => {
   const testUtils = require('../../test/utils');
@@ -172,80 +170,5 @@ describe('user: bookings controller', () => {
     expect(plugins[0].createBooking).toHaveBeenCalled();
     expect(plugins[0].createBooking.mock.calls[0][0].payload).toEqual(payload);
     expect(plugins[0].createBooking.mock.calls[0][0].token).toEqual(token);
-  });
-
-  describe('bookingsProductSearch caching', () => {
-    it('should not return empty products when TTR expires and stale cache exists, even if refresh yields empty products', async () => {
-      const cacheModule = require('../../cache'); // Get the mocked cache module
-
-      const initialProductsInCache = [{ productId: 'cached1', name: 'Cached Product One', optionId: 'opt1' }];
-      const productsFromPluginRefresh = []; // Simulate plugin returning empty on refresh
-
-      // Configure a short TTR for this test via plugin's cacheSettings
-      // This relies on token.ttlForProducts being undefined for this test user's token setup
-      const ttrInSeconds = 1;
-      plugins[0].cacheSettings = { bookingsProductSearch: { ttr: ttrInSeconds } };
-
-      const expectedCacheKeyForProducts = hash({
-        userId,
-        hint: newIntegration.tokenHint, // 'testingToken'
-        operationId: 'bookingsProductSearch',
-      });
-      const lastUpdatedKey = `${expectedCacheKeyForProducts}:lastUpdated`;
-      const lockKey = `${expectedCacheKeyForProducts}:lock`;
-
-      cacheModule.get.mockImplementation(async ({ pluginName, key }) => {
-        // Ensure calls are for the correct plugin
-        if (pluginName !== appKey) return null;
-
-        if (key === expectedCacheKeyForProducts) {
-          return { products: initialProductsInCache };
-        }
-        if (key === lastUpdatedKey) {
-          // Simulate that lastUpdated is older than TTR
-          return Date.now() - (ttrInSeconds * 1000 * 2);
-        }
-        if (key === lockKey) {
-          return null; // No lock exists
-        }
-        return null;
-      });
-
-      // Mock the plugin's underlying product search method
-      // (app.searchProducts || app.searchProductsForItinerary)
-      // Assuming searchProducts is the relevant method for 'travelgate' plugin
-      plugins[0].searchProducts.mockResolvedValue({ products: productsFromPluginRefresh });
-
-      const searchPayload = { searchInput: 'test' };
-      const { products: resultProducts } = await doApiPost({
-        url: `/bookings/${appKey}/${userId}/${newIntegration.tokenHint}/products/search`,
-        token: userToken,
-        payload: searchPayload,
-      });
-
-      // **** This is the key assertion based on the user's requirement ****
-      // It expects the stale cache (initialProductsInCache) rather than productsFromPluginRefresh
-      // This assertion will likely FAIL with the current code if it returns the empty refreshed data.
-      expect(resultProducts).toEqual(initialProductsInCache);
-      expect(resultProducts.length).toBeGreaterThan(0);
-
-
-      // Verify interactions
-      expect(cacheModule.get).toHaveBeenCalledWith(expect.objectContaining({ pluginName: appKey, key: expectedCacheKeyForProducts }));
-      expect(cacheModule.get).toHaveBeenCalledWith(expect.objectContaining({ pluginName: appKey, key: lastUpdatedKey }));
-      expect(cacheModule.get).toHaveBeenCalledWith(expect.objectContaining({ pluginName: appKey, key: lockKey }));
-
-      expect(plugins[0].searchProducts).toHaveBeenCalledTimes(1); // Refresh was attempted
-
-      // Verify cache operations for locking and updating
-      expect(cacheModule.save).toHaveBeenCalledWith(expect.objectContaining({ pluginName: appKey, key: lockKey, value: true }));
-      expect(cacheModule.save).toHaveBeenCalledWith(expect.objectContaining({
-        pluginName: appKey,
-        key: expectedCacheKeyForProducts,
-        value: { products: productsFromPluginRefresh }, // Cache updated with new (empty) data
-      }));
-      expect(cacheModule.save).toHaveBeenCalledWith(expect.objectContaining({ pluginName: appKey, key: lastUpdatedKey })); // lastUpdated timestamp was updated
-      expect(cacheModule.drop).toHaveBeenCalledWith(expect.objectContaining({ pluginName: appKey, key: lockKey })); // Lock was released
-    });
   });
 });
